work around Enterprise strangeness
GitHub Enterprise uses a sentinel value of 999999999999 to denote
"unlimited".

+++ Octokit/Models/Plan.cs
@@ -8,7 +8,8 @@ namespace Octokit
         /// <summary>
         /// The number of collaborators allowed with this plan.
         /// </summary>
-        public int Collaborators { get; set; }
+        /// <remarks>This returns <see cref="long"/> because GitHub Enterprise uses a sentinel value of 999999999999 to denote an "unlimited" number of collaborators.</remarks>
+        public long Collaborators { get; set; }
 
         /// <summary>
         /// The name of the plan.
@@ -18,12 +19,14 @@ namespace Octokit
         /// <summary>
         /// The number of private repositories allowed with this plan.
         /// </summary>
-        public int PrivateRepos { get; set; }
+        /// <remarks>This returns <see cref="long"/> because GitHub Enterprise uses a sentinel value of 999999999999 to denote an "unlimited" number of plans.</remarks>
+        public long PrivateRepos { get; set; }
 
         /// <summary>
         /// The amount of disk space allowed with this plan.
         /// </summary>
-        public int Space { get; set; }
+        /// <remarks>This returns <see cref="long"/> because GitHub Enterprise uses a sentinel value of 999999999999 to denote an "unlimited" amount of disk space.</remarks>
+        public long Space { get; set; }
 
         /// <summary>
         /// The billing email for the organization. Only has a value in response to editing an organization.
